app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
boto3==1.34.144
requests==2.32.3
python-dotenv==1.0.1
